+++ api-scanner/src/test/resources/annotated-interface/java/net/corda/example/DoNotImplementAnnotation.java
@@ -1,7 +1,7 @@
 package net.corda.example;
 
-import net.corda.annotation.*;
-import net.corda.core.*;
+import net.corda.annotation.AnAnnotation;
+import net.corda.v5.base.annotations.DoNotImplement;
 
 @AnAnnotation
 @DoNotImplement
