+++ api-scanner/src/test/resources/annotated-interface/java/net/corda/example/HasInheritedAnnotation.java
@@ -1,7 +1,8 @@
 package net.corda.example;
 
-import net.corda.annotation.*;
-import net.corda.core.*;
+import net.corda.annotation.AlsoInherited;
+import net.corda.annotation.IsInherited;
+import net.corda.annotation.NotInherited;
 
 @NotInherited
 @IsInherited
