+++ jar-filter/build.gradle
@@ -102,8 +102,10 @@ tasks.withType(Test).configureEach {
         languageVersion = of(11)
     }
 
-    // Configures the @RequiresKotlin14 annotation.
-    systemProperty 'test.kotlin.api', test_kotlin_api_version
+    doFirst {
+        // Configures the @RequiresKotlin14 annotation.
+        systemProperty 'test.kotlin.api', test_kotlin_api_version
+    }
 
     // Resolve warnings about "illegal reflective access" that
     // are generated by Gradle's own Groovy library.
